Start WFM refactor
